requirements.txt
matplotlib
seaborn
beautifulsoup4
schedule
pymysql
pytz
requests
pyyaml
lxml